Don't call sleep_for(), instead use the Qt socket method waitForReadyRead() per azaparov's suggestion.
Simplify the approach a great deal : just expect data
points and plot them. In my application, I needed
a layer between the data generator and PlotJuggler,
so I started adding that layer in this plugin. But
azaparov suggested keeping the layer separate, which
turns out to be much nicer in my application, and
makes this plug-in simpler